refactor: UserLeagueHistory 도메인 아키텍처 개선
- infrastructure 패키지 제거, repository 패키지 추가
