+++ src/main/java/gravit/code/season/batch/SeasonBatchService.java
@@ -9,7 +9,7 @@ import gravit.code.season.domain.Season;
 import gravit.code.season.repository.SeasonRepository;
 import gravit.code.season.service.port.SeasonClosedCache;
 import gravit.code.userLeague.domain.UserLeagueRepository;
-import gravit.code.userLeagueHistory.infrastructure.UserLeagueHistoryRepository;
+import gravit.code.userLeagueHistory.repository.UserLeagueHistoryRepository;
 import java.sql.SQLException;
 import java.time.Clock;
 import java.time.LocalDateTime;
