+++ src/test/java/gravit/code/season/batch/SeasonBatchServiceTest.java
@@ -8,7 +8,7 @@
 //import gravit.code.season.repository.SeasonRepository;
 //import gravit.code.season.service.port.SeasonClosedCache;
 //import gravit.code.userLeague.domain.UserLeagueRepository;
-//import gravit.code.userLeagueHistory.infrastructure.UserLeagueHistoryRepository;
+//import gravit.code.userLeagueHistory.repository.UserLeagueHistoryRepository;
 //import org.junit.jupiter.api.Test;
 //import org.junit.jupiter.api.extension.ExtendWith;
 //import org.mockito.ArgumentCaptor;
